+++ packages/bvaughn-architecture-demo/components/console/ConsoleRoot.tsx
