+++ packages/bvaughn-architecture-demo/components/console/index.tsx
@@ -1,3 +1,3 @@
-import Console from "./Console";
+import ConsoleRoot from "./ConsoleRoot";
 
-export default Console;
+export default ConsoleRoot;
